requirements.txt
lightning==2.4.0
matplotlib==3.10.0
networkx==3.3
numpy==2.2.2
pandas==2.2.3
Pillow==11.1.0
rdkit==2024.3.3
torch==2.3.1
selfies==2.2.0

# Additional requirement for SAFE notation package
loguru==0.7.3